Set default course date to Djangos "today"
This removes the two warnings that pop up during system self check and
avoids that every time someone is making a migration to the database the
date.today() function is being evaluated and the current date is being
mirated into the db for the course begin and end defaults.
Requires one last migration, though.