Fix `IntegrityError` in registration, close #86
The problem was related to the fact that the forms for user information
and student information are not connected (i.e., connected through
inheritance) but their models in the backend are. Thus, the user
information is not transferred into the student information model when
the user provides a zih login and a faculty.

Fixed by working around the problem ¯\_(ツ)_/¯